requirements.txt
termcolor
hydra-core
mne
einops
scikit-learn
h5py
pandas
lxml
cmake
# dlib
mediapipe
opencv-python
wandb
gdown
torch
natsort